+++ pyproject.toml
@@ -15,6 +15,7 @@ datetime = "^5.5"
 prometheus-client = "^0.20.0"
 pytest = "^8.2.0"
 pytest-cov = "^5.0.0"
+gevent = "^24.2.1"
 
 
 [build-system]
